Move temp file function
